unified alfa 1 with 1.0.1-a2

+++ app/Http/Middleware/Super.php
@@ -16,7 +16,7 @@ class Super
      * @return mixed
      */
     public function handle(Request $request, Closure $next){
-        if( (int)Auth::user()->is_super !== 1 ){
+        if(Auth::user()->is_super !== 1){
             return response()->view('platform.super-check');
         }
         return $next($request);
